requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.0.0
numpy>=1.21.0
torch>=1.9.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

